requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.2
django-allauth==0.50.0
django-braces==1.15.0
django-cors-headers==3.11.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.0.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
waitress==2.0.0
whitenoise==6.0.0
wrapt==1.14.1
